pyproject.toml
[project]
name = "quadsci_challenge"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11,<3.12"

dependencies = [
    "dowhy==0.12",
    "graphviz==0.20.3",
    "tqdm>=4.66.5",
    "pandas==2.0.3",
    "pandas-stubs==2.2.1.240316",
    "pytest-xdist==3.6.1",
    "pytest-cov==6.0.0",
    "pydot==3.0.4",
    "types-tqdm==4.66.0.20240106",
    "types-requests==2.32.0.20241016 ",
    "typing_extensions==4.9.0",
    "typing-inspect==0.9.0",
    "ipython==8.28.0 ",
    "ipykernel==6.29.5",
    "ipywidgets==8.1.5",
    "mlflow==2.21.3",
    "notebook==7.2.2",
    "seaborn==0.13.2",
    "scikit-learn==1.3.0",
    "scipy==1.11.1",
    "shap==0.47.1",
    "statsmodels==0.14.4",
    "torch-geometric==2.6.1",
    "torch==2.0.1 ; platform_machine != 'x86_64'",
    "torch==2.0.1+cu118 ; platform_machine == 'x86_64'",
    "torchvision==0.15.2+cu118; platform_machine == 'x86_64'",
    "torchvision==0.15.2; platform_machine != 'x86_64'",
    "xgboost==3.0.0",
]
[project.optional-dependencies]
format = [
    "autoflake==2.3.0",
    "isort==5.13.2",
    "black==24.2.0",
    "docformatter==1.7.5",
    "mypy==1.8.0",
    "pydocstyle==6.3.0",
    "coverage==7.6.8",
]


# Index configuration for torch and torchvision with gpu
[tool.uv.sources]
torch = { index = "torch-cu118" }
torchvision = { index = "torch-cu118" }

[[tool.uv.index]]
name = "torch-cu118"
url = "https://download.pytorch.org/whl/cu118"
explicit = true
#[tool.pytest.ini_options]
#addopts = "--cov=measurements --cov-report=term-missing --cov-report=xml:coverage-results/coverage.xml"